Check if port is occupied before trying to release it

+++ packages/gulp-tasks/tasks/watch.js
@@ -6,6 +6,7 @@ const remember = require('gulp-remember');
 const watch = require('gulp-watch');
 const path = require('path');
 const fs = require('fs');
+const net = require('net');
 const { promisify } = require('util');
 const exec = promisify(require('child_process').exec);
 
@@ -14,7 +15,7 @@ const sharedState = require('../gulpState.js');
 exports.__requiresConfig = true;
 
 exports.default = gulpConfig => {
-  const { files, occupiedPorts } = gulpConfig;
+  const { files, occupiedPorts } = gulpConfig;
 
   function watchTask() {
     let hotReloadedCacheKeys = [];
@@ -95,20 +96,50 @@ exports.default = gulpConfig => {
 
   function checkAndReleasePorts() {
     const occupants = Object.keys(occupiedPorts);
-    
-    gutil.log(`Releasing ports occupied by ${occupants.join(', ')}`);
 
     return Promise.all(occupants.map(occupant => {
       const port = occupiedPorts[occupant];
 
-      const command = process.platform === 'win32'
-        ? `Stop-Process -Id (Get-NetTCPConnection -LocalPort ${port}).OwningProcess -Force`
-        : `lsof -i:${port} | grep LISTEN | awk '{print $2}' | xargs kill -9`;
+      return isPortOccupied(port)
+        .then(occupied => {
+          if (!occupied) {
+            return;
+          }
 
-      return exec(command).catch(() => {
-        throw Error(`Unable to free port ${port} occupied by ${occupant}. Try freeing this port manually.`);
-      });
+          gutil.log(`Releasing port occupied by ${occupant}.`);
+
+          const command = process.platform === 'win32'
+            ? `Stop-Process -Id (Get-NetTCPConnection -LocalPort ${port}).OwningProcess -Force`
+            : `lsof -i:${port} | grep LISTEN | awk '{print $2}' | xargs kill -9`;
+
+          return exec(command).catch(() => null);
+        })
+        .catch(error => {
+          throw Error(`Unable to determine if port ${port} is occupied.`);
+        });
     }));
+  }
+
+  function isPortOccupied(port) {
+
+    return new Promise((resolve, reject) => {
+      const tester = net.createServer();
+
+      tester.once('error', error => {
+        if (error.code !== 'EADDRINUSE') {
+          return reject(error);
+        }
+        resolve(true);
+      });
+
+      tester.once('listening', () => {
+        tester
+          .once('close', () => resolve(false))
+          .close()
+      });
+
+      tester.listen(port);
+    });
   }
 
   return {
